MIM-156 拆分工作区胶囊布局组件
